src/web-app/pyproject.toml
[project]
name = "context-aware-vision-grounded-kb-agent"
version = "0.1.0"
description = "Context Aware & Vision Grounded KB Agent — conversational agent with structure-preserving chunking and image-aware search over Azure AI Search"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "openai>=1.60.0",
    "azure-cosmos>=4.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["integration: requires running services and Azure credentials"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
prerelease = "allow"
